+++ game/graphics/opengl_renderer/Shader.cpp
@@ -81,4 +81,5 @@ ShaderLibrary::ShaderLibrary() {
   at(ShaderId::OCEAN_TEXTURE) = {"ocean_texture"};
   at(ShaderId::OCEAN_TEXTURE_MIPMAP) = {"ocean_texture_mipmap"};
   at(ShaderId::OCEAN_COMMON) = {"ocean_common"};
+  at(ShaderId::SHADOW) = {"shadow"};
 }
