+++ game/graphics/opengl_renderer/shaders/shadow.frag
@@ -1,0 +1,8 @@
+#version 430 core
+
+out vec4 color;
+uniform vec4 color_uniform;
+
+void main() {
+    color = color_uniform;
+}
